tools/requirements-optional.txt
# ===== Operational Security Specialization Dependencies =====


#--- Core Monitoring Dependencies ---
# Collect application metrics for Prometheus
prometheus-client==0.20.0

#--- Structured Logging (Recommended) ---
# JSON-formatted logging for better auditability
python-json-logger==2.0.7

#--- Security Simulation Tools (Optional) ---
# Generate mock data for incident response and testing
faker==24.8.0
requests==2.31.0

#--- Advanced Monitoring (Optional) ---
# System resources monitoring 
psutil>=5.9.0                      
# Memory usage analysis
pympler==1.0.1
# Network interface info
netifaces>=0.11.0

#--- Visualization Tools (Development Only) ---
# Optional for producing analysis plots and charts, system architecture diagrams
numpy==1.24.3
matplotlib==3.7.1
pandas==1.5.3
jupyter==1.0.0
graphviz==0.20.1

#YAML parsing and validation
pyyaml>=6.0

#--- Security Testing & Analysis ---
# For security testing and validation
bandit==1.7.5
safety==2.3.5
jq >= 1.6
